move daysLeftToFix to common. Use it to figure out which vulnerabilities to include
