+++ packages/common/src/functions.test.ts
@@ -2,6 +2,7 @@ import type { SNSEvent } from 'aws-lambda';
 import {
 	anghammaradThreadKey,
 	branchProtectionCtas,
+	daysLeftToFix,
 	getEnvOrThrow,
 	getGithubAppSecret,
 	parseEvent,
@@ -11,7 +12,7 @@ import {
 	toNonEmptyArray,
 	topicMonitoringProductionTagCtas,
 } from './functions';
-import type { NonEmptyArray } from './types';
+import type { NonEmptyArray, RepocopVulnerability } from './types';
 
 function isValidUrl(str: string) {
 	try {
@@ -201,6 +202,57 @@ describe('stringToSeverity', () => {
 		expect(stringToSeverity('medium')).toBe('medium');
 		expect(stringToSeverity('high')).toBe('high');
 		expect(stringToSeverity('critical')).toBe('critical');
+	});
+});
+
+describe('daysLeftToFix', () => {
+	const veryOldVuln: RepocopVulnerability = {
+		source: 'Dependabot',
+		full_name: 'guardian/repo',
+		open: true,
+		severity: 'high',
+		package: 'leftpad',
+		urls: ['example.com'],
+		ecosystem: 'pip',
+		alert_issue_date: new Date('2021-01-01'),
+		is_patchable: true,
+		cves: ['CVE-123'],
+	};
+	test('should return 0 if we exceed the SLA', () => {
+		expect(daysLeftToFix(veryOldVuln)).toBe(0);
+	});
+	test('should return 30 if a high vuln was raised in the last 24 hours', () => {
+		function hoursAgo(hours: number): Date {
+			const date = new Date();
+			date.setHours(date.getHours() - hours);
+			return date;
+		}
+
+		const oneHourOld: RepocopVulnerability = {
+			...veryOldVuln,
+			alert_issue_date: hoursAgo(1),
+		};
+
+		const twentyThreeHoursOld = {
+			...oneHourOld,
+			alert_issue_date: hoursAgo(23),
+		};
+
+		const twentyFiveHoursOld = {
+			...oneHourOld,
+			alert_issue_date: hoursAgo(25),
+		};
+		expect(daysLeftToFix(oneHourOld)).toBe(30);
+		expect(daysLeftToFix(twentyThreeHoursOld)).toBe(30);
+		expect(daysLeftToFix(twentyFiveHoursOld)).toBe(29);
+	});
+	test('should return 2 if a critical vuln was raised today', () => {
+		const newCriticalVuln: RepocopVulnerability = {
+			...veryOldVuln,
+			severity: 'critical',
+			alert_issue_date: new Date(),
+		};
+		expect(daysLeftToFix(newCriticalVuln)).toBe(2);
 	});
 });
